requirements.txt
algoliasearch-django
django
djangorestframework
djangorestframework-simplejwt
pyyaml
pillow
requests
django-cors-headers
python-decouple